Object File Seperation
