--- Canvas/src/dataTypes.js
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=dataTypes.js.map
