+++ Canvas/ts/api.ts
@@ -1,0 +1,7 @@
+export interface MaximumValues {
+
+}
+
+export interface DataYearThumbnail {
+    
+}
