--- Canvas/ts/dataTypes.ts
@@ -1,3 +1,0 @@
-export type DataYearThumbnail = {
-    
-}
