+++ Canvas/ts/manager.ts
@@ -1,0 +1,19 @@
+import { Controller } from './controller.js'
+
+export class Manager {
+    controller: Controller;
+
+
+
+    constructor(_controller: Controller) {
+        this.controller = _controller;
+        this.controller.managerChangeCallback = this.contextChangeCallback;
+    }
+
+    contextChangeCallback(_active: string) {
+        console.log(_active)
+    }
+}
+
+
+
