+++ Canvas/ts/sketch.ts
@@ -1,5 +1,6 @@
-import { Country, ContourObject, HoverAnimation, ColourScheme, Controller, RGB } from './types.js'
+import { Country, ContourObject, HoverAnimation, ColourScheme, RGB } from './types.js'
 import { DEFAULTFILEPATHS, DEFAULTOFFSETS, COLOURSLIGHT, SCALINGCONSTANTS } from './config.js';
+import { Controller } from './controller.js'
 
 export async function setupCanvas(_colorScheme: ColourScheme, _labelContainer: JQuery<HTMLElement>): Promise<Controller> {
     let ctryList: Map<string, Country>;
@@ -116,7 +117,7 @@ export async function setupCanvas(_colorScheme: ColourScheme, _labelContainer: J
             scaleCountries();
             positionCountries();
             
-            controller.updateCanvasAttributes(scaleFactorX, scaleFactorY)
+            controller.updateCanvasAttributes(scaleFactorX, scaleFactorY, widthTranslation, heightTranslation)
             controller.positionLabels();
             controller.renderLabels();
         }
@@ -211,5 +212,5 @@ export async function setupCanvas(_colorScheme: ColourScheme, _labelContainer: J
         p5Context = new p5(sketch),
         timeout(50)
     ])
-    return controller = new Controller({ctryList, staticObjList}, _labelContainer, heightTranslation, widthTranslation, scaleFactorX, scaleFactorY)
+    return controller = new Controller({ctryList, staticObjList}, _labelContainer, scaleFactorX, scaleFactorY, widthTranslation, heightTranslation)
 }
