+++ Canvas/ts/types.ts
@@ -1,5 +1,5 @@
 import { DEFAULTFILEPATHS, COLOURSLIGHT, SCALINGCONSTANTS, ViewTypes, createLabel, createSelectItems } from "./config.js";
-
+import { Controller } from './controller.js' 
 
 export interface ImgPaths {
     readonly rainfall: string,
@@ -38,215 +38,6 @@ export interface ColourScheme {
     gradientNull: RGB,
     labelBackground: string,
     labelText: string
-}
-
-
-
-// Implementation of https://stackoverflow.com/a/5624139
-// export function RGBFromHex(_hex: string): RGB | null {
-//     var result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(_hex);
-//     return result ? {
-//       r: parseInt(result[1], 16),
-//       g: parseInt(result[2], 16),
-//       b: parseInt(result[3], 16)
-//     } : null;
-// }
-
-export class Manager {
-    controller: Controller;
-
-
-
-    constructor(_controller: Controller) {
-        this.controller = _controller;
-        this.controller.managerChangeCallback = this.contextChangeCallback;
-    }
-
-    contextChangeCallback(_active: string) {
-        console.log(_active)
-    }
-}
-
-export class Controller {
-    countryList: Map<string, Country>;
-    staticObjectList: Map<string, ContourObject>
-    labelContainer: JQuery<HTMLElement>;
-    labelList: Map<string, Label>
-    lastActive: string;
-
-    canvasHeightTranslation: number;
-    canvasWidthTranslation: number;
-    canvasScaleFactorX: number;
-    canvasScaleFactorY: number;
-
-    viewActive: number;
-
-    managerChangeCallback: (_active: string) => void;
-
-    activeChangeCallback(_active: string) {
-        if (_active == this.lastActive) return;
-        this.lastActive = _active;
-        // pass message to manager.
-        this.managerChangeCallback(_active)
-    }
-
-    handleHoverAnimations(_ctry: Country) {
-        let shade = 12;  
-
-        if (_ctry.hover == true && _ctry.active == false) {
-            if(_ctry.animationMap.has('hoverReverse')) {
-                // shade = shade - Math.abs(_ctry.animationMap.get('hoverReverse').endAnimation());
-                _ctry.animationMap.delete('hoverReverse')
-            }
-            // console.log(shade)
-            if (!_ctry.animationMap.has('hover')) {
-                let startFill = {
-                    r: _ctry.rootFill.r - shade,
-                    g: _ctry.rootFill.g - shade,
-                    b: _ctry.rootFill.b - shade
-                }
-                let animation = new HoverAnimation(_ctry,'hover',255,startFill,shade)
-                _ctry.animationMap.set('hover', animation);
-            }
-        } else {
-            if(_ctry.animationMap.has('hover')) {
-                shade = -_ctry.animationMap.get('hover').endAnimation();
-                _ctry.animationMap.delete('hover')
-
-
-                if (!_ctry.animationMap.has('hoverReverse')) {
-                    let startFill = {
-                        r: _ctry.rootFill.r + shade,
-                        g: _ctry.rootFill.g + shade,
-                        b: _ctry.rootFill.b + shade
-                    }
-                    let animation = new HoverAnimation(_ctry,'hoverReverse',255,startFill,shade)
-                    _ctry.animationMap.set('hoverReverse', animation);
-                }
-            }
-        }
-    }
-
-    viewChange(_view: number) {
-        if(this.viewActive == _view) return;
-        this.viewActive = _view;
-        this.renderSelectBarLabels();
-    }
-
-    checkLabelOwnerExists(_labelName: string): boolean {
-        for (let key of this.countryList.keys()) {
-            if (_labelName == key) return true;
-        }
-        return false;
-    }
-
-    setupLabels(_labelConfig: object): void {
-        for (let config in Object.entries(_labelConfig)) {
-            let info = Object.values(_labelConfig)[config]
-            if (!this.checkLabelOwnerExists(info.owner)) continue;
-            
-            let element: JQuery<HTMLElement> = createLabel(COLOURSLIGHT)
-            let country: Country = this.countryList.get(info.owner);
-
-            element.find('#canvas-label-title').text(info.title)
-
-            let label: Label = {
-                content: element,
-                offsetX: info.offset[0],
-                offsetY: info.offset[1]
-            }
-
-            label.content.on('mouseenter', (e: JQuery.Event) => {
-                country.hoverLock = true;
-                country.hover = true;
-                this.handleHoverAnimations(country)
-            }) 
-            
-            label.content.on('mouseleave', (e: JQuery.Event) => {
-                country.hoverLock = false;
-            })
-
-            label.content.on('click', (e: JQuery.Event) => {
-                country.active = true;
-            })
-
-            this.labelList.set(info.owner, label)
-        }
-    }
-
-    updateCanvasAttributes(_scaleFactorX: number, _scaleFactorY: number) {
-        this.canvasScaleFactorX = _scaleFactorX;
-        this.canvasScaleFactorY = _scaleFactorY;
-    }
-
-
-    positionLabels(): void {
-        for (let [key, label] of this.labelList.entries()) {
-            if (!this.checkLabelOwnerExists(key)) continue;
-
-            label.content.css({
-                // "top": (this.labelContainer.innerHeight() * 0.7) + (label.offsetY * this.canvasScaleFactorY),
-                // "left": (this.labelContainer.innerWidth() * 0.65) + (label.offsetX * this.canvasScaleFactorX)
-                "top": this.canvasHeightTranslation + (label.offsetY * this.canvasScaleFactorY),
-                "left": this.canvasWidthTranslation + (label.offsetX * this.canvasScaleFactorX)
-            })
-        }
-
-    }
-
-    renderLabels() {
-        this.labelList.forEach((obj, key) => {
-            obj.content.appendTo(this.labelContainer)
-        })
-
-
-    }
-    
-    renderSelectBarLabels() {
-        let selectBar = this.labelContainer.find('#canvas-label-select-bar')
-        if(!selectBar) return;
-        selectBar.empty()
-        for(let key = 0; key < Object.entries(ViewTypes).length; key++) {
-            let element: JQuery<HTMLElement> = createSelectItems(Object.entries(ViewTypes)[key][0], COLOURSLIGHT);            
-            
-            if(key == this.viewActive) {
-                element.addClass("canvas-label-select-bar-item-active")
-            } else {
-                element.on('click', (e: JQuery.Event) => {
-                    this.viewChange(key)
-                })
-            }
-            
-            element.appendTo(selectBar);
-        }
-    }
-
-    constructor(_objects: { ctryList: Map<string, Country>, staticObjList: Map<string, ContourObject> },
-        _labelContainer: JQuery<HTMLElement>,
-        _canvasHeightTranslation: number,
-        _canvasWidthTranslation: number,
-        _scaleFactorX: number,
-        _scaleFactorY: number
-    ) {
-        this.countryList = _objects.ctryList;
-        this.staticObjectList = _objects.staticObjList;
-        
-        this.labelContainer = _labelContainer;
-        this.labelList = new Map();
-        this.lastActive = undefined;
-
-        this.canvasHeightTranslation = _canvasHeightTranslation;
-        this.canvasWidthTranslation = _canvasWidthTranslation;
-        this.updateCanvasAttributes(_scaleFactorX, _scaleFactorY);
-
-        this.viewActive = ViewTypes.highTemp
-    }
-}
-
-type Label = {
-    content: JQuery<HTMLElement>,
-    offsetX: number,
-    offsetY: number
 }
 
 export type RGB = {
